refactored custom subscriptions into their own module
